GP-0: Moving addExternalBlock() to a more appropriate class
