+++ Ghidra/Features/Base/src/main/java/ghidra/app/util/bin/format/macho/commands/chained/DyldChainedFixups.java
@@ -19,12 +19,12 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+import ghidra.app.util.MemoryBlockUtils;
 import ghidra.app.util.bin.BinaryReader;
 import ghidra.app.util.bin.format.macho.dyld.DyldChainedPtr;
 import ghidra.app.util.bin.format.macho.dyld.DyldChainedPtr.DyldChainType;
 import ghidra.app.util.bin.format.macho.dyld.DyldFixup;
 import ghidra.app.util.importer.MessageLog;
-import ghidra.app.util.opinion.AbstractProgramLoader;
 import ghidra.app.util.opinion.MachoProgramBuilder;
 import ghidra.program.model.address.Address;
 import ghidra.program.model.listing.*;
@@ -153,7 +153,7 @@ public class DyldChainedFixups {
 					.mapToLong(DyldFixup::size)
 					.sum();
 			if (externalSize > 0) {
-				extAddr = AbstractProgramLoader.addExternalBlock(program, externalSize, log);
+				extAddr = MemoryBlockUtils.addExternalBlock(program, externalSize, log);
 			}
 		}
 		catch (Exception e) {
